Added support for apps using namespaces
For coherency, I had to place the class into a namespace (not included in this fix of course). 

Prior to the fix, class_exists('MongoClient') would fail, as MongoClient was not in the same namespace, hence the change to class_exists('\MongoClient'). This fix works with both namespace in use or not, so no harm, no foul.